data mining
Definition of Hierarchical Clustering
A hierarchical clustering approach is based on the determination of successive clusters based on previously defined clusters. It's a technique aimed more toward grouping data into a tree of clusters called dendrograms, which graphically represents the hierarchical relationship between the underlying clusters.

Comparison of Hierarchical Clustering to Other Clustering Techniques
Hierarchical clustering is a powerful algorithm, but it is not the only one out there, and each type of clustering comes with its set of advantages and drawbacks. 

Let’s understand how it compares to other types of clustering, such as K-means and model-based clustering. Many more techniques exist, but these two, in addition to hierarchical clustering, are widely used and provide a framework to help better understand the others. 

You can learn more about clustering in machine learning in our separate article, covering five essential clustering algorithms. 

Hierarchical clustering vs K Means clustering
Unlike Hierarchical clustering, K-means clustering seeks to partition the original data points into “K” groups or clusters where the user specifies “K” in advance. 

The general idea is to look for clusters that minimize the squared Euclidean distance of all the points from the centers over all attributes (variables or features) and merge those individuals in an iterative manner. 

Our K-means Clustering in Python with Scikit-learn tutorial will help you understand the inner workings of K-means clustering with an interesting case study.

Benefits 
It is computationally efficient compared to hierarchical clustering and can be used to analyze large data sets. 
K-means is easier to understand and implement. 
Drawbacks
It is less flexible than hierarchical clustering because it constrains the user to specify the number of clusters beforehand, which may not be obvious in some situations.  
The result is not stable and changes from one iteration to another for the same data set.
It is more sensitive to outliers because the use of outliers in the data impacts the mean of the cluster. 
Both k-means and hierarchical clustering are incapable of handling categorical data directly and may not work well with data that is not continuous or has very large variance.
Despite its limitations, k-means clustering is still a popular method due to its ease of use and computational efficiency. It is frequently used as a reference point for comparing the performance of other clustering techniques. 

Model-based clustering
Both K-means and hierarchical clustering techniques use a distance matrix to represent the distances between all the points in the dataset. Model-based clustering, on the other hand, applies statistical techniques to identify clusters in the data. Below is the general process: 

Decide the statistical model to use and choose the number of clusters. 
Fit the model on the data. 
Identify the clusters based on the model parameters.
Benefits 
Model-based clustering is more flexible than hierarchical clustering because it allows the use of different models to identify different types of clusters. 
It works better on data with complex shapes or structures.
Drawbacks
It’s computationally more expensive than hierarchical clustering, especially for large data. 
It requires a better understanding of statistical modeling techniques since the choice of the model can affect the final result. 
Similarly to K-means clustering, it requires the number of clusters to be specified beforehand. 
Applications of Hierarchical Clustering
Hierarchical clustering has a variety of applications in our day-to-day life, including (but by no means limited to) biology, image processing, marketing, economics, and social network analysis.